Split up RecipeDetails into seperate files
