+++ app/src/main/java/com/omelan/cofi/pages/details/Dialogs.kt
@@ -1,0 +1,55 @@
+package com.omelan.cofi.pages.details
+
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.platform.LocalClipboardManager
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.AnnotatedString
+import com.omelan.cofi.R
+import com.omelan.cofi.appDeepLinkUrl
+import kotlinx.coroutines.launch
+
+@Composable
+fun rememberCopyAutomateLink(snackbarState: SnackbarHostState, recipeId: Int): () -> Unit {
+    val clipboardManager = LocalClipboardManager.current
+    val coroutineScope = rememberCoroutineScope()
+    val snackbarMessage = stringResource(id = R.string.snackbar_copied)
+
+    clipboardManager.setText(AnnotatedString(text = "$appDeepLinkUrl/recipe/$recipeId"))
+    return {
+        coroutineScope.launch {
+            snackbarState.showSnackbar(message = snackbarMessage)
+        }
+    }
+}
+
+@Composable
+fun DirectLinkDialog(dismiss: () -> Unit, onConfirm: () -> Unit) {
+    AlertDialog(
+        onDismissRequest = dismiss,
+        confirmButton = {
+            TextButton(onClick = onConfirm) {
+                Text(text = stringResource(id = R.string.button_copy))
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = dismiss) {
+                Text(text = stringResource(id = R.string.button_cancel))
+            }
+        },
+        icon = {
+            Icon(
+                painter = painterResource(id = R.drawable.ic_link),
+                contentDescription = null,
+            )
+        },
+        title = {
+            Text(text = stringResource(R.string.recipe_details_automation_dialog_title))
+        },
+        text = {
+            Text(text = stringResource(R.string.recipe_details_automation_dialog_text))
+        },
+    )
+}
