+++ app/src/main/java/com/omelan/cofi/pages/details/RatioBottomSheet.kt
@@ -1,0 +1,7 @@
+package com.omelan.cofi.pages.details
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun RatioBottomSheet(timeMultiplier: Double, weightMultiplier: Double) {
+}
